backend/requirements.txt
paho-mqtt>=2.0
requests>=2.31
PyYAML>=6.0
